feat: new set and set id breakdown in components
