+++ web/src/app/admin/documents/sets/new/page.tsx
@@ -2,63 +2,10 @@
 
 import { AdminPageTitle } from "@/components/adminPageComponents/Title";
 import { BookmarkIcon } from "@/components/icons/icons";
-import { DocumentSetCreationForm } from "@/components/adminPageComponents/documents/AdminDocumentsSetsCreationForm";
-import {
-  useConnectorCredentialIndexingStatus,
-  useUserGroups,
-} from "@/lib/hooks";
-import { ThreeDotsLoader } from "@/components/Loading";
-import { usePopup } from "@/hooks/common/usePopup";
-import { Card } from "@tremor/react";
 import { BackButton } from "@/components/BackButton";
-import { ErrorCallout } from "@/components/ErrorCallout";
-import { useRouter } from "next/navigation";
-import { refreshDocumentSets } from "@/hooks/admin/documents/useDocumentsSets";
+import DocumentsNewSetCreationFormHOC from "@/components/adminPageComponents/documents/AdminDocumentsNewSetCreationFormHOC";
 
-function Main() {
-  const { popup, setPopup } = usePopup();
-  const router = useRouter();
 
-  const {
-    data: ccPairs,
-    isLoading: isCCPairsLoading,
-    error: ccPairsError,
-  } = useConnectorCredentialIndexingStatus();
-
-  // EE only
-  const { data: userGroups, isLoading: userGroupsIsLoading } = useUserGroups();
-
-  if (isCCPairsLoading || userGroupsIsLoading) {
-    return <ThreeDotsLoader />;
-  }
-
-  if (ccPairsError || !ccPairs) {
-    return (
-      <ErrorCallout
-        errorTitle="Failed to fetch Connectors"
-        errorMsg={ccPairsError}
-      />
-    );
-  }
-
-  return (
-    <>
-      {popup}
-
-      <Card>
-        <DocumentSetCreationForm
-          ccPairs={ccPairs}
-          userGroups={userGroups}
-          onClose={() => {
-            refreshDocumentSets();
-            router.push("/admin/documents/sets");
-          }}
-          setPopup={setPopup}
-        />
-      </Card>
-    </>
-  );
-}
 
 const Page = () => {
   return (
@@ -70,7 +17,7 @@ const Page = () => {
         title="New Document Set"
       />
 
-      <Main />
+      <DocumentsNewSetCreationFormHOC />
     </div>
   );
 };
